Added indents and removed useless == TRUE
in an if statement, when the output of the test is a logical, no need
to do if(test == TRUE). In this case it doubles execution time: it runs
the test, outputs a logical, and check if this logical is TRUE and if
all that is TRUE, does the if. It can be simplified to just if(test)
where it asses simply if(TRUE) and then does the if. Same for the
!=TRUE. You can replace them by if(!test).